refactor: decompose 4 more monolith files into submodules (issue #2486 sprint 10)
- motion_optimization.py (788 → 276 LOC) → _motion_opt_types, _motion_opt_trajectory,
  _motion_opt_simulation, _motion_primitive_library
- topography.py (774 → 40 LOC) → _topography_types, _topography_data,
  _topography_factory, _topography_io
- geometric_primitives.py (774 → 29 LOC) → _primitives_base, _primitive_shapes,
  _distance_queries
- kinematic_forces.py (772 → 82 LOC) → _kfa_core, _kfa_forces, _kfa_analysis,
  _kfa_effective_mass

All public symbols remain importable from original module paths.
