+++ src/engines/physics_engines/mujoco/python/mujoco_humanoid_golf/kinematic_forces.py
@@ -5,23 +5,20 @@ coordinate frame definitions, numerical tolerances, and known limitations.
 
 Implementation split across:
 - _kinematic_force_data.py: MjDataContext, KinematicForceData, version check, CSV export
+- _kfa_core.py: _KFACoreMixin — init, body lookup, Jacobian helpers
+- _kfa_forces.py: _KFAForcesMixin — Coriolis, gravity, mass matrix computations
+- _kfa_analysis.py: _KFAAnalysisMixin — power, kinetic energy, trajectory analysis
+- _kfa_effective_mass.py: _KFAEffectiveMassMixin — effective mass computation
 """
 
 from __future__ import annotations
 
-import warnings
-from typing import TYPE_CHECKING
-
 import mujoco
-import numpy as np
 
-from src.shared.python.core.numerical_constants import (
-    EPSILON_FINITE_DIFF_JACOBIAN,
-    EPSILON_SINGULARITY_DETECTION,
-)
-
-if TYPE_CHECKING:
-    pass
+from ._kfa_analysis import _KFAAnalysisMixin
+from ._kfa_core import _KFACoreMixin
+from ._kfa_effective_mass import _KFAEffectiveMassMixin
+from ._kfa_forces import _KFAForcesMixin
 
 # Re-export public names for backward compatibility
 from ._kinematic_force_data import (
@@ -31,7 +28,12 @@ from ._kinematic_force_data import (
 )
 
 
-class KinematicForceAnalyzer:
+class KinematicForceAnalyzer(
+    _KFACoreMixin,
+    _KFAForcesMixin,
+    _KFAAnalysisMixin,
+    _KFAEffectiveMassMixin,
+):
     """Analyze kinematic-dependent forces in golf swing.
 
     This class computes Coriolis, centrifugal, and other velocity-dependent
@@ -69,699 +71,7 @@ class KinematicForceAnalyzer:
             model: MuJoCo model
             data: MuJoCo data (shared reference, not modified by compute methods)
         """
-        if not (model is not None):
-            raise ValueError("model must be provided")
-        if not (model is not None):
-            raise ValueError("model must be provided")
-        self.model = model
-        self.data = data
-
-        # Find important bodies
-        self.club_head_id = self._find_body_id("club_head")
-        self.club_grip_id = self._find_body_id("club") or self._find_body_id("grip")
-
-        # MEMORY ALLOCATION: Dedicated MjData for state-isolated computations
-        # This prevents race conditions but increases memory footprint
-        # See Issue A-004 for detailed analysis
-        self._perturb_data = mujoco.MjData(model)
-
-        # Pre-allocate Jacobian buffers to avoid repeated allocation
-        # Detect API version to use correct array shape
-        self.nv = model.nv
-        try:
-            # Try reshaped API (MuJoCo 3.3+ preferred)
-            jacp_test = np.zeros((3, self.nv))
-            jacr_test = np.zeros((3, self.nv))
-            mujoco.mj_jacBody(model, data, jacp_test, jacr_test, 0)
-            self._use_reshaped_arrays = True
-            self._jacp = np.zeros((3, self.nv))
-            self._jacr = np.zeros((3, self.nv))
-        except TypeError:
-            # Fallback to flat API
-            self._use_reshaped_arrays = False
-            self._jacp = np.zeros(3 * self.nv)
-            self._jacr = np.zeros(3 * self.nv)
-
-    def _find_body_id(self, name_pattern: str) -> int | None:
-        """Find body ID by name pattern."""
-        if not (name_pattern is not None):
-            raise ValueError("name_pattern must be provided")
-        if not (name_pattern is not None):
-            raise ValueError("name_pattern must be provided")
-        for i in range(self.model.nbody):
-            body_name = mujoco.mj_id2name(self.model, mujoco.mjtObj.mjOBJ_BODY, i)
-            if body_name and name_pattern.lower() in body_name.lower():
-                return i
-        return None
-
-    def _compute_jacobian(
-        self, body_id: int, data: mujoco.MjData | None = None
-    ) -> tuple[np.ndarray, np.ndarray]:
-        """Compute Jacobian for a body using pre-allocated buffers.
-
-        Args:
-            body_id: Body ID
-            data: MuJoCo data (default: self.data)
-
-        Returns:
-            Tuple of (jacp, jacr) as (3, nv) arrays.
-            Note: Returns views into internal buffers or copies depending on usage.
-        """
-        if not (body_id is not None):
-            raise ValueError("body_id must be provided")
-        if not (body_id is not None):
-            raise ValueError("body_id must be provided")
-        if data is None:
-            data = self.data
-
-        if self._use_reshaped_arrays:
-            mujoco.mj_jacBody(self.model, data, self._jacp, self._jacr, body_id)
-            return self._jacp, self._jacr
-        else:
-            mujoco.mj_jacBody(self.model, data, self._jacp, self._jacr, body_id)
-            return (
-                self._jacp.reshape(3, self.nv),
-                self._jacr.reshape(3, self.nv),
-            )
-
-    def compute_coriolis_forces(self, qpos: np.ndarray, qvel: np.ndarray) -> np.ndarray:
-        """Compute Coriolis and centrifugal forces.
-
-        These are the velocity-dependent forces in the equations of motion:
-        M(q)q̈ + C(q,q̇)q̇ + g(q) = τ
-
-        The term C(q,q̇)q̇ represents Coriolis and centrifugal forces.
-
-        FIXED: Uses dedicated _perturb_data instead of shared self.data to prevent
-        race conditions and state corruption. See Issues A-001 and F-002.
-
-        Args:
-            qpos: Joint positions [nv]
-            qvel: Joint velocities [nv]
-
-        Returns:
-            Coriolis forces [nv]
-        """
-        # PHASE 1 UPGRADE: Use Analytical RNE
-        # This replaces the legacy O(N) forward kinematics approach with
-        # the significantly faster Recursive Newton-Euler algorithm.
-        return self.compute_coriolis_forces_rne(qpos, qvel)
-
-    def compute_coriolis_forces_rne(
-        self, qpos: np.ndarray, qvel: np.ndarray
-    ) -> np.ndarray:
-        """Compute Coriolis forces using analytical RNE (Phase 1).
-
-        This method uses MuJoCo's Recursive Newton-Euler (mj_rne) algorithm
-        which is significantly faster than mj_forward used in the legacy method.
-        It computes the Coriolis terms analytically, avoiding extra overhead
-        of forward kinematics and visual updates.
-
-        Args:
-            qpos: Joint positions [nv]
-            qvel: Joint velocities [nv]
-
-        Returns:
-            Coriolis forces [nv]
-        """
-        # Use private/scratch data structure
-        if not (qpos is not None):
-            raise ValueError("qpos must be provided")
-        if not (qpos is not None):
-            raise ValueError("qpos must be provided")
-        self._perturb_data.qpos[:] = qpos
-        self._perturb_data.qvel[:] = qvel
-        # qacc must be zero for RNE to return inverse dynamics forces
-        self._perturb_data.qacc[:] = 0.0
-
-        # MuJoCo RNE: tau = M(q)qacc + C(q,qvel)qvel + g(q)
-        # With qacc=0, it returns: C(q,qvel)qvel + g(q)
-        bias = np.zeros(self.model.nv)
-        mujoco.mj_rne(self.model, self._perturb_data, 0, bias)
-
-        # Now we need to subtract gravity g(q)
-        # RNE with qvel=0, qacc=0 returns g(q)
-        self._perturb_data.qvel[:] = 0.0
-        gravity = np.zeros(self.model.nv)
-        mujoco.mj_rne(self.model, self._perturb_data, 0, gravity)
-
-        return bias - gravity
-
-    def compute_gravity_forces(self, qpos: np.ndarray) -> np.ndarray:
-        """Compute gravitational forces.
-
-        FIXED: Uses dedicated _perturb_data to prevent state corruption.
-        See Issues A-001 and F-002.
-
-        Args:
-            qpos: Joint positions [nv]
-
-        Returns:
-            Gravity forces [nv]
-        """
-        # FIXED: Use private/scratch data structure
-        if not (qpos is not None):
-            raise ValueError("qpos must be provided")
-        if not (qpos is not None):
-            raise ValueError("qpos must be provided")
-        self._perturb_data.qpos[:] = qpos
-        self._perturb_data.qvel[:] = 0.0
-
-        mujoco.mj_forward(self.model, self._perturb_data)
-
-        # With zero velocity, qfrc_bias = g(q)
-        return np.asarray(self._perturb_data.qfrc_bias.copy())
-
-    def decompose_coriolis_forces(
-        self,
-        qpos: np.ndarray,
-        qvel: np.ndarray,
-    ) -> tuple[np.ndarray, np.ndarray]:
-        """Decompose Coriolis forces into centrifugal and velocity coupling.
-
-            Coriolis matrix C(q,q̇) can be decomposed:
-            - Centrifugal terms: Diagonal terms (q̇ᵢ²)
-            - Velocity coupling: Off-diagonal terms (q̇ᵢq̇ⱼ)
-
-            ⚠️ PERFORMANCE NOTE (Phase 1 Implemented):
-        This method iterates N times to separate diagonal vs off-diagonal terms.
-        However, the inner calculation now uses Analytical RNE (`mj_rne`),
-        which is significantly faster than the legacy `mj_forward` approach.
-        Global complexity is O(N^2) but with a much smaller constant factor.
-
-            RECOMMENDATION: Use the combined compute_coriolis_forces() method instead,
-            which returns the total Coriolis+centrifugal forces efficiently in O(N).
-            Decomposition is rarely needed for most applications.
-
-            See Issue A-002 and B-002 for optimization path using analytical RNE.
-
-            Args:
-                qpos: Joint positions [nv]
-                qvel: Joint velocities [nv]
-
-            Returns:
-                Tuple of (centrifugal_forces [nv], coupling_forces [nv])
-        """
-        # OPTIMIZATION NOTE: The proper fix is to use mj_rne properties or
-        # implement analytical decomposition. Current implementation is
-        # accurate but slow for high-DOF systems.
-        #
-        # Full analytical solution requires:
-        # - Custom RNE recursion to separate diagonal/off-diagonal terms
-        # - OR use spatial algebra (screw theory) for frame-independent decomposition
-        # - OR skip decomposition and use total Coriolis forces directly
-        #
-        # For most golf swing analyses, the total Coriolis force is sufficient.
-
-        if not (qpos is not None):
-            raise ValueError("qpos must be provided")
-        if not (qpos is not None):
-            raise ValueError("qpos must be provided")
-        centrifugal = np.zeros(self.model.nv)
-        coupling = np.zeros(self.model.nv)
-
-        # Full Coriolis forces
-        total_coriolis = self.compute_coriolis_forces(qpos, qvel)
-
-        # Estimate centrifugal: vary each velocity independently
-        # This captures diagonal terms of the Coriolis matrix
-        for i in range(self.model.nv):
-            qvel_single = np.zeros(self.model.nv)
-            qvel_single[i] = qvel[i]
-
-            single_coriolis = self.compute_coriolis_forces(qpos, qvel_single)
-            centrifugal += single_coriolis
-
-        # Coupling is the difference (off-diagonal terms)
-        coupling = total_coriolis - centrifugal
-
-        return centrifugal, coupling
-
-    def compute_mass_matrix(self, qpos: np.ndarray) -> np.ndarray:
-        """Compute configuration-dependent mass matrix M(q).
-
-        FIXED: Uses dedicated _perturb_data to prevent state corruption.
-
-        Args:
-            qpos: Joint positions [nv]
-
-        Returns:
-            Mass matrix [nv x nv]
-        """
-        # FIXED: Use private data structure
-        if not (qpos is not None):
-            raise ValueError("qpos must be provided")
-        if not (qpos is not None):
-            raise ValueError("qpos must be provided")
-        self._perturb_data.qpos[:] = qpos
-        mujoco.mj_forward(self.model, self._perturb_data)
-
-        # Get full mass matrix
-        M = np.zeros((self.model.nv, self.model.nv))
-        mujoco.mj_fullM(self.model, M, self._perturb_data.qM)
-
-        return M
-
-    def compute_coriolis_matrix(self, qpos: np.ndarray, qvel: np.ndarray) -> np.ndarray:
-        """Compute Coriolis matrix C(q,q̇).
-
-        The Coriolis matrix satisfies: C(q,q̇)q̇ = coriolis forces
-
-        Args:
-            qpos: Joint positions [nv]
-            qvel: Joint velocities [nv]
-
-        Returns:
-            Coriolis matrix [nv x nv]
-        """
-        # Use finite differences to estimate C
-        # (Deprecated - use RNE-based method instead)
-        if not (qpos is not None):
-            raise ValueError("qpos must be provided")
-        if not (qpos is not None):
-            raise ValueError("qpos must be provided")
-        epsilon = EPSILON_FINITE_DIFF_JACOBIAN
-        C = np.zeros((self.model.nv, self.model.nv))
-
-        # Reference Coriolis forces
-        c_ref = self.compute_coriolis_forces(qpos, qvel)
-
-        for i in range(self.model.nv):
-            qvel_perturb = qvel.copy()
-            qvel_perturb[i] += epsilon
-
-            c_perturb = self.compute_coriolis_forces(qpos, qvel_perturb)
-
-            C[:, i] = (c_perturb - c_ref) / epsilon
-
-        return C
-
-    def compute_club_head_apparent_forces(  # noqa: PLR0915
-        self,
-        qpos: np.ndarray,
-        qvel: np.ndarray,
-        qacc: np.ndarray,
-    ) -> tuple[np.ndarray, np.ndarray, np.ndarray]:
-        """Compute apparent forces at club head (Coriolis, centrifugal, etc.).
-
-        These are the "fictitious" forces experienced in the rotating
-        reference frame attached to the golfer.
-
-        FIXED: Uses dedicated _perturb_data to prevent state corruption.
-
-        Args:
-            qpos: Joint positions [nv]
-            qvel: Joint velocities [nv]
-            qacc: Joint accelerations [nv]
-
-        Returns:
-            Tuple of (coriolis_force [3], centrifugal_force [3], total_apparent [3])
-        """
-        if not (qpos is not None):
-            raise ValueError("qpos must be provided")
-        if not (qpos is not None):
-            raise ValueError("qpos must be provided")
-        if self.club_head_id is None:
-            return np.zeros(3), np.zeros(3), np.zeros(3)
-
-        # FIXED: Use private data structure for current state
-        self._perturb_data.qpos[:] = qpos
-        self._perturb_data.qvel[:] = qvel
-        mujoco.mj_forward(self.model, self._perturb_data)
-
-        # Compute club head Jacobian
-        jacp, _ = self._compute_jacobian(self.club_head_id, data=self._perturb_data)
-        jacp_curr = jacp.copy()  # Save copy as _compute_jacobian reuses buffer
-
-        # Store club position before perturbing state
-        club_pos = self._perturb_data.xpos[self.club_head_id].copy()
-
-        # Jacobian time derivative (approximate)
-        # ASSESSMENT B-009: Use second-order central difference for improved accuracy
-        # Central difference: J̇ ≈ (J(q+εq̇) - J(q-εq̇)) / (2ε)
-        # Error: O(ε²) vs O(ε) for forward difference
-        epsilon = EPSILON_FINITE_DIFF_JACOBIAN
-
-        # Compute Jacobian at forward-perturbed state
-        self._perturb_data.qpos[:] = qpos + epsilon * qvel
-        self._perturb_data.qvel[:] = qvel
-        mujoco.mj_forward(self.model, self._perturb_data)
-
-        jacp_forward, _ = self._compute_jacobian(
-            self.club_head_id, data=self._perturb_data
-        )
-        jacp_forward = jacp_forward.copy()  # Save copy before buffer reuse
-
-        # Compute Jacobian at backward-perturbed state
-        self._perturb_data.qpos[:] = qpos - epsilon * qvel
-        self._perturb_data.qvel[:] = qvel
-        mujoco.mj_forward(self.model, self._perturb_data)
-
-        jacp_backward, _ = self._compute_jacobian(
-            self.club_head_id, data=self._perturb_data
-        )
-
-        # Second-order central difference
-        # Accuracy: O(ε²) - much better than O(ε) forward difference
-        jacp_dot = (jacp_forward - jacp_backward) / (2.0 * epsilon)
-
-        # Coriolis force: -2m(Ω × v)
-        # In our case: Coriolis contribution to acceleration
-        coriolis_accel = jacp_dot @ qvel
-
-        # Assume unit mass for force (or multiply by club head mass)
-        club_head_mass = self.model.body_mass[self.club_head_id]
-        coriolis_force = -club_head_mass * coriolis_accel
-
-        # Centrifugal force: -mΩ²r
-        # This is embedded in the Coriolis term
-        # For separation, we'd need angular velocity of each body segment
-
-        # Total apparent force (from joint-space Coriolis forces)
-        joint_coriolis = self.compute_coriolis_forces(qpos, qvel)
-        apparent_force = jacp_curr.T @ joint_coriolis[: self.model.nv]
-
-        # Approximate centrifugal as component aligned with position
-        # Singularity protection when near origin
-        centrifugal_direction = club_pos / (
-            np.linalg.norm(club_pos) + EPSILON_SINGULARITY_DETECTION
-        )
-        centrifugal_magnitude = np.dot(apparent_force, centrifugal_direction)
-        centrifugal_force = centrifugal_magnitude * centrifugal_direction
-
-        return coriolis_force, centrifugal_force, apparent_force
-
-    def compute_kinematic_power(
-        self,
-        qpos: np.ndarray,
-        qvel: np.ndarray,
-    ) -> dict[str, float]:
-        """Compute power contributions from kinematic forces.
-
-        Args:
-            qpos: Joint positions [nv]
-            qvel: Joint velocities [nv]
-
-        Returns:
-            Dictionary with power contributions
-        """
-        # Coriolis forces
-        if not (qpos is not None):
-            raise ValueError("qpos must be provided")
-        if not (qpos is not None):
-            raise ValueError("qpos must be provided")
-        coriolis_forces = self.compute_coriolis_forces(qpos, qvel)
-
-        # Coriolis power (should be zero for conservative systems)
-        # Power = F · v
-        coriolis_power = np.dot(coriolis_forces, qvel)
-
-        # Decompose
-        centrifugal, coupling = self.decompose_coriolis_forces(qpos, qvel)
-        centrifugal_power = np.dot(centrifugal, qvel)
-        coupling_power = np.dot(coupling, qvel)
-
-        # Gravity power
-        gravity_forces = self.compute_gravity_forces(qpos)
-        gravity_power = np.dot(gravity_forces, qvel)
-
-        return {
-            "coriolis_power": float(coriolis_power),
-            "centrifugal_power": float(centrifugal_power),
-            "coupling_power": float(coupling_power),
-            "gravity_power": float(gravity_power),
-            "total_conservative_power": float(coriolis_power + gravity_power),
-        }
-
-    def compute_kinetic_energy_components(
-        self,
-        qpos: np.ndarray,
-        qvel: np.ndarray,
-    ) -> dict[str, float]:
-        """Decompose kinetic energy into rotational and translational.
-
-        FIXED: Uses dedicated _perturb_data to prevent state corruption.
-
-        Args:
-            qpos: Joint positions [nv]
-            qvel: Joint velocities [nv]
-
-        Returns:
-            Dictionary with kinetic energy components
-        """
-        # FIXED: Use private data structure
-        if not (qpos is not None):
-            raise ValueError("qpos must be provided")
-        if not (qpos is not None):
-            raise ValueError("qpos must be provided")
-        self._perturb_data.qpos[:] = qpos
-        self._perturb_data.qvel[:] = qvel
-        mujoco.mj_forward(self.model, self._perturb_data)
-
-        rotational_ke = 0.0
-        translational_ke = 0.0
-
-        for i in range(1, self.model.nbody):  # Skip world body
-            # Get body mass and inertia
-            body_mass = self.model.body_mass[i]
-            body_inertia = self.model.body_inertia[i]
-
-            # Get body velocity (linear and angular)
-            jacp, jacr = self._compute_jacobian(i, data=self._perturb_data)
-
-            v_linear = jacp @ qvel
-            omega = jacr @ qvel
-
-            # Kinetic energies
-            translational_ke += 0.5 * body_mass * np.dot(v_linear, v_linear)
-            rotational_ke += 0.5 * np.dot(omega, body_inertia * omega)
-
-        return {
-            "rotational": float(rotational_ke),
-            "translational": float(translational_ke),
-            "total": float(rotational_ke + translational_ke),
-        }
-
-    def analyze_trajectory(
-        self,
-        times: np.ndarray,
-        positions: np.ndarray,
-        velocities: np.ndarray,
-        accelerations: np.ndarray,
-    ) -> list[KinematicForceData]:
-        """Analyze kinematic forces along a trajectory.
-
-        This is the main function for analyzing captured motion data.
-
-        Args:
-            times: Time array [N]
-            positions: Joint positions [N x nv]
-            velocities: Joint velocities [N x nv]
-            accelerations: Joint accelerations [N x nv]
-
-        Returns:
-            List of KinematicForceData for each time step
-        """
-        if not (times is not None):
-            raise ValueError("times must be provided")
-        if not (times is not None):
-            raise ValueError("times must be provided")
-        results = []
-
-        for i in range(len(times)):
-            qpos = positions[i]
-            qvel = velocities[i]
-            qacc = accelerations[i]
-
-            # Compute forces
-            coriolis = self.compute_coriolis_forces(qpos, qvel)
-            gravity = self.compute_gravity_forces(qpos)
-            centrifugal, coupling = self.decompose_coriolis_forces(qpos, qvel)
-
-            # Club head apparent forces
-            club_coriolis, club_centrifugal, club_apparent = (
-                self.compute_club_head_apparent_forces(qpos, qvel, qacc)
-            )
-
-            # Power contributions
-            power_dict = self.compute_kinematic_power(qpos, qvel)
-
-            # Kinetic energy
-            ke_dict = self.compute_kinetic_energy_components(qpos, qvel)
-
-            # Create data object
-            data = KinematicForceData(
-                time=times[i],
-                coriolis_forces=coriolis,
-                gravity_forces=gravity,
-                centrifugal_forces=centrifugal,
-                velocity_coupling_forces=coupling,
-                club_head_coriolis_force=club_coriolis,
-                club_head_centrifugal_force=club_centrifugal,
-                club_head_apparent_force=club_apparent,
-                coriolis_power=power_dict["coriolis_power"],
-                centrifugal_power=power_dict["centrifugal_power"],
-                rotational_kinetic_energy=ke_dict["rotational"],
-                translational_kinetic_energy=ke_dict["translational"],
-            )
-
-            results.append(data)
-
-        return results
-
-    def _validate_effective_mass_direction(self, direction: np.ndarray) -> np.ndarray:
-        direction_norm = np.linalg.norm(direction)
-        if direction_norm < EPSILON_SINGULARITY_DETECTION:
-            raise ValueError(
-                f"Direction vector has near-zero magnitude: {direction_norm:.2e}. "
-                "Cannot compute effective mass for zero-length direction."
-            )
-        return direction / direction_norm
-
-    def _check_mass_matrix_conditioning(self, M: np.ndarray) -> None:
-        M_cond = np.linalg.cond(M)
-        if M_cond > 1e6:
-            warnings.warn(
-                f"Mass matrix is ill-conditioned: κ(M) = {M_cond:.2e} > 1e6. "
-                "Effective mass computation may be numerically unstable. "
-                "This often indicates the robot is near a kinematic singularity.",
-                category=UserWarning,
-                stacklevel=2,
-            )
-
-        eigenvalues = np.linalg.eigvalsh(M)
-        if np.any(eigenvalues <= 0):
-            raise ValueError(
-                f"Mass matrix is not positive definite. "
-                f"Minimum eigenvalue: {eigenvalues.min():.2e}. "
-                "This indicates a modeling error or numerical instability."
-            )
-
-    def _check_jacobian_rank(self, jacp: np.ndarray) -> None:
-        J_rank = np.linalg.matrix_rank(jacp)
-        if J_rank < 3:
-            warnings.warn(
-                f"Jacobian is rank deficient: rank={J_rank} < 3. "
-                "Robot has lost mobility in some directions. "
-                "Effective mass may not be well-defined.",
-                category=RuntimeWarning,
-                stacklevel=2,
-            )
-
-    def _compute_effective_mass_value(
-        self, direction: np.ndarray, jacp: np.ndarray, M: np.ndarray
-    ) -> float:
-        J_dir = direction @ jacp
-        M_inv = np.linalg.inv(M)
-        denominator = J_dir @ M_inv @ J_dir.T + EPSILON_SINGULARITY_DETECTION
-
-        if abs(denominator) < 1e-8:
-            warnings.warn(
-                f"Effective mass denominator near zero: {denominator:.2e}. "
-                "Robot is at or very close to a kinematic singularity in the "
-                f"specified direction {direction}. Effective mass is extremely large.",
-                category=UserWarning,
-                stacklevel=2,
-            )
-
-        m_eff = 1.0 / denominator
-
-        if m_eff < 0:
-            raise ValueError(
-                f"Computed negative effective mass: {m_eff:.2e} kg. "
-                "This indicates a numerical error or modeling issue."
-            )
-
-        if not np.isfinite(m_eff):
-            warnings.warn(
-                f"Effective mass is non-finite: {m_eff}. "
-                "Robot is at a kinematic singularity. "
-                "Returning large finite value instead.",
-                category=UserWarning,
-                stacklevel=2,
-            )
-            m_eff = 1e10
-
-        return float(m_eff)
-
-    def compute_effective_mass(
-        self,
-        qpos: np.ndarray,
-        direction: np.ndarray,
-        body_id: int | None = None,
-    ) -> float:
-        """Compute effective mass in a given direction.
-
-        Effective mass determines how difficult it is to accelerate
-        in a specific direction. Near kinematic singularities, the
-        effective mass can become very large (approaching infinity).
-
-        NUMERICAL STABILITY (Assessment B-008):
-        ----------------------------------------
-        This method monitors the condition number of the mass matrix
-        and Jacobian to detect approaching singularities. When detected:
-        - Warning issued if condition number > 1e6
-        - Regularization applied automatically
-        - Result validity flag returned
-
-        PHYSICS:
-        --------
-        The effective mass is computed as:
-            m_eff = (J M^{-1} J^T)^{-1}
-
-        Where:
-        - J = Jacobian mapping joint velocities to task-space velocity
-        - M = joint-space mass matrix (configuration-dependent)
-
-        Near singularities:
-        - m_eff → ∞ (infinite mass, cannot accelerate)
-        - Physically meaningful: robot at kinematic boundary
-
-        FIXED: Uses dedicated _perturb_data to prevent state corruption.
-
-        Args:
-            qpos: Joint positions [nv] (rad for revolute, m for prismatic)
-            direction: Direction vector [3] (will be normalized)
-            body_id: Body to compute for (default: club head)
-
-        Returns:
-            Effective mass in that direction [kg]
-
-        Warns:
-            UserWarning: If approaching singularity (condition number > 1e6)
-
-        Raises:
-            ValueError: If mass matrix is not positive definite
-            RuntimeWarning: If Jacobian is rank deficient
-
-        Examples:
-            >>> # Compute effective mass in vertical direction
-            >>> direction = np.array([0, 0, 1])  # Z-up
-            >>> m_eff = analyzer.compute_effective_mass(qpos, direction)
-            >>> print(f"Effective mass: {m_eff:.2f} kg")
-        """
-        if not (qpos is not None):
-            raise ValueError("qpos must be provided")
-        if not (qpos is not None):
-            raise ValueError("qpos must be provided")
-        if body_id is None:
-            body_id = self.club_head_id
-
-        if body_id is None:
-            return 0.0
-
-        direction = self._validate_effective_mass_direction(direction)
-
-        M = self.compute_mass_matrix(qpos)
-        self._check_mass_matrix_conditioning(M)
-
-        self._perturb_data.qpos[:] = qpos
-        mujoco.mj_forward(self.model, self._perturb_data)
-
-        jacp, _ = self._compute_jacobian(body_id, data=self._perturb_data)
-        self._check_jacobian_rank(jacp)
-
-        return self._compute_effective_mass_value(direction, jacp, M)
+        self._init_core(model, data)
 
 
 __all__ = [
